updating MVC

+++ app/Http/Controllers/ArticleController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Article;
-
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+// use Illuminate\Auth\Access\Gate;
 
 class ArticleController extends Controller
 {
@@ -49,6 +51,7 @@ class ArticleController extends Controller
       $article->title = request()->title;
       $article->body = request()->body;
       $article->category_id = request()->category_id;
+      $article ->user_id = auth()->user()->id;
       $article->save();
 
       return redirect('/articles');
@@ -56,8 +59,20 @@ class ArticleController extends Controller
 
    public function delete($id){
       $article = Article::find($id);
-      $article->delete();
 
-      return redirect('/articles')->with('info', 'Article Deleted');
+      if( Gate::allows('article-delete', $article)){
+         $article->delete();
+         return redirect('/articles')->with('info', 'Article Deleted');
+      }
+      else{
+         return redirect('/articles')->with('error', 'Unauthorize');
+     }
+     
+   }
+
+   // for authorization
+   public function __construct()
+   {
+      $this->middleware('auth')->except(['index', 'detail']);
    }
 }
