+++ app/Http/Controllers/CommentController.php
@@ -4,21 +4,35 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
     public function create(){
         $comment = new Comment();
         $comment ->content = request()->content;
         $comment ->article_id = request()->article_id;
+        // $comment ->user_id = Auth::id();
+        $comment ->user_id = auth()->user()->id;
         $comment->save();
         
         return back();
     }
     public function delete($id) {
         $comment = Comment::find($id);
-        $comment->delete();
 
-        return back();
+        if(Gate::allows('comment-delete', $comment)){
+            $comment->delete();
+            return back();
+        }else{
+            return back()->with('error', 'Unauthorize');
+        }
+
     }
+
 }
